API BREAK: libxdot: use size_t instead of int for op sizes/counts
This allows exceeding operation counts of 2³¹ - 1, which is not particularly
likely. But its more important effect is to make arithmetic operations less
error prone and make it impossible by-construction to arrive at a negative
operation size or count.

Doing this as a single sweeping change seemed less error prone and clearer,
despite how large the resulting diff is.

For lines that would have been changed anyway, this commit also fixes white
space and tightens variable scoping where possible.

+++ lib/xdot/xdot.h
@@ -10,6 +10,7 @@
 
 #pragma once
 
+#include <stddef.h>
 #include <stdio.h>
 
 #ifdef __cplusplus
@@ -143,15 +144,15 @@ struct _xdot_op {
 #define XDOT_PARSE_ERROR 1
 
 typedef struct {
-    int cnt;  /* no. of xdot ops */
-    int sz;   /* sizeof structure containing xdot_op as first field */
+    size_t cnt;  /* no. of xdot ops */
+    size_t sz;   /* sizeof structure containing xdot_op as first field */
     xdot_op* ops;
     freefunc_t freefunc;
     int flags;
 } xdot;
 
 typedef struct {
-    int cnt;  /* no. of xdot ops */
+    size_t cnt;  /* no. of xdot ops */
     int n_ellipse;
     int n_polygon;
     int n_polygon_pts;
@@ -169,8 +170,8 @@ typedef struct {
 } xdot_stats;
 
 /* ops are indexed by xop_kind */
-XDOT_API xdot* parseXDotF (char*, drawfunc_t opfns[], int sz);
-XDOT_API xdot* parseXDotFOn (char*, drawfunc_t opfns[], int sz, xdot*);
+XDOT_API xdot *parseXDotF(char*, drawfunc_t opfns[], size_t sz);
+XDOT_API xdot *parseXDotFOn(char*, drawfunc_t opfns[], size_t sz, xdot*);
 XDOT_API xdot* parseXDot (char*);
 XDOT_API char* sprintXDot (xdot*);
 XDOT_API void fprintXDot (FILE*, xdot*);
